Completed all the components with their interfaces and params shifted
