+++ src/components/atoms/Tag/index.ts
@@ -1,0 +1,2 @@
+export * from './tag'
+export * from './tag.types'
