+++ src/components/atoms/TextInput/index.ts
@@ -1,0 +1,2 @@
+export * from './TextInput'
+export * from './TextInput.types'
