+++ src/components/atoms/Typography/index.ts
@@ -1,1 +1,2 @@
 export * from './Typography';
+export * from './Typography.types'
