+++ src/components/atoms/UserProfile/index.ts
@@ -1,0 +1,2 @@
+export * from './UserProfile'
+export * from './UserProfile.types'
