+++ src/components/atoms/table/index.ts
@@ -1,0 +1,2 @@
+export * from './table'
+export * from './table.types'
