+++ src/components/atoms/upload/index.ts
@@ -1,0 +1,2 @@
+export * from './upload'
+export * from './upload.types'
