+++ src/components/atoms/uploadItem/UploadItem.types.ts
@@ -1,0 +1,12 @@
+
+/**
+ * @param label - Label for the component
+ * @param progress - Progress of the component
+ * @param size - Size of the file
+ */
+
+export interface UploadItemsProps {
+  label: string;
+  progress: number;
+  size: string;
+}
