+++ src/components/atoms/uploadItem/index.ts
@@ -1,0 +1,2 @@
+export * from './UploadItem'
+export * from './UploadItem.types'
