+++ src/helpers/extend-colors-storybook.tsx
@@ -16,7 +16,7 @@ import { EmptySection, EmptySectionProps } from '~/components/atoms/EmptySection
 import { Pagination, PaginationProps } from '~/components/atoms/Pagination/pagination';
 import { Table, TableProps } from '~/components/atoms/table/table';
 import { Tag, TagProps } from '~/components/atoms/Tag/tag';
-import { UserProfile, UserProfileProps } from '~/components/atoms/UserProfile/user-profile';
+import { UserProfile, UserProfileProps } from '~/components/atoms/UserProfile/UserProfile';
 import { AmountInputField, AmountInputProps } from '~/components/atoms/AmountInput';
 import { Calendar, CalendarProps } from '~/components/atoms/calendar/calendar';
 import { Checkbox, CheckboxProps } from '~/components/atoms/checkbox';
@@ -38,8 +38,8 @@ import { Radio, RadioProps } from '~/components/atoms/Radio/Radio';
 import { SearchInputField, SearchInputProps } from '~/components/atoms/SearchInput/SearchInput';
 import { SelectInputField, SelectInputProps } from '~/components/atoms/SelectInput/SelectInput';
 import { SiteInputField, SiteInputProps } from '~/components/atoms/SiteInput/SiteInput';
-import { TextInputField, TextInputProps } from '~/components/atoms/TextInput/text-input';
-import { UploadItem, UploadItemsProps } from '~/components/atoms/uploadItem/upload-item';
+import { TextInputField, TextInputProps } from '~/components/atoms/TextInput/TextInput';
+import { UploadItem, UploadItemsProps } from '~/components/atoms/uploadItem/UploadItem';
 import { Upload, UploadProps } from '~/components/atoms/upload/upload';
 import { ButtonGroup, ButtonGroupProps } from '~/components/atoms/ButtonGroup';
 import { ButtonWrapper } from '~/components/atoms/ButtonWrapper/ButtonWrapper';
