+++ src/stories/dropdown.stories.tsx
@@ -1,7 +1,7 @@
 import type { ComponentStory, ComponentMeta } from '@storybook/react';
 import { Button } from '../components/atoms/Button/Button';
 import { Item } from '../components/atoms/sidebar/menu';
-import { UserProfile } from '../components/atoms/UserProfile/user-profile';
+import { UserProfile } from '../components/atoms/UserProfile/UserProfile';
 import { Dropdown } from '../components/atoms/Dropdown/dropdown';
 import React from 'react';
