+++ src/stories/text-input.stories.tsx
@@ -1,6 +1,6 @@
 import type { ComponentStory, ComponentMeta } from '@storybook/react';
 import React from 'react';
-import { TextInputField } from '../components/atoms/TextInput/text-input';
+import { TextInputField } from '../components/atoms/TextInput/TextInput';
 import { ComponentWithColors, ComponentName } from '~/helpers/extend-colors-storybook';
 import { extendTheme, base, ColorShade } from '~/themes';
